Rename package from coderpad_api to coderpad and simplify client class names
Rename the package directory from `coderpad_api` to `coderpad` and the client
classes from `CoderPadClient`/`AsyncCoderPadClient` to `CoderPad`/`AsyncCoderPad`.

Usage is now `from coderpad.client import CoderPad`.
