+++ docs/source/conf.py
@@ -69,10 +69,10 @@ intersphinx_mapping = {
 }
 nitpicky = True
 nitpick_ignore: list[tuple[str, str]] = [
-    ("py:class", "coderpad_api.types._T"),
+    ("py:class", "coderpad.types._T"),
 ]
 nitpick_ignore_regex = [
-    (r"py:class", r"coderpad_api\._dict_types\..*"),
+    (r"py:class", r"coderpad\._dict_types\..*"),
 ]
 warning_is_error = True
